検索の画面下部にボタン設置、チャットへ飛びます
