--- lib/feature/chat/chat.dart
@@ -1,276 +1,0 @@
-import 'dart:async';
-import 'dart:io';
-
-import 'package:cloud_firestore/cloud_firestore.dart';
-import 'package:dash_chat/dash_chat.dart';
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:firebase_storage/firebase_storage.dart' as firebase_storage;
-import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
-import 'package:intl/intl.dart';
-import 'package:uuid/uuid.dart';
-
-class MyHomePage extends StatefulWidget {
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-  TextEditingController _editingController = TextEditingController();
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Dash Chat"),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(8.0),
-        child: Form(
-          key: _formKey,
-          child: Column(
-            children: <Widget>[
-              TextFormField(
-                controller: _editingController,
-                decoration: InputDecoration(hintText: "Enter Name here..."),
-                validator: (val) {
-                  if (val.length < 3) {
-                    return "Name to short";
-                  }
-                  return null;
-                },
-              ),
-              SizedBox(
-                height: 40.0,
-              ),
-              FlatButton(
-                child: Text("Next"),
-                onPressed: () async {
-                  if (_formKey.currentState.validate()) {
-                    print(_editingController.value.text);
-                    try {
-                      await FirebaseAuth.instance.signInAnonymously();
-
-                      Navigator.of(context).push(
-                        MaterialPageRoute(
-                          builder: (context) => ChatScreen(
-                            username: _editingController.text,
-                            uuid: Uuid().v4().toString(),
-                          ),
-                        ),
-                      );
-                    } catch (e) {
-                      print(e);
-                    }
-                  }
-                },
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class ChatScreen extends StatefulWidget {
-  final String username;
-  final String uuid;
-
-  ChatScreen({this.username, this.uuid});
-
-  @override
-  _ChatScreenState createState() => _ChatScreenState();
-}
-
-class _ChatScreenState extends State<ChatScreen> {
-  ChatUser user = ChatUser();
-  final GlobalKey<DashChatState> _chatViewKey = GlobalKey<DashChatState>();
-  List<ChatMessage> messages = List<ChatMessage>();
-  var m = List<ChatMessage>();
-
-  var i = 0;
-
-  @override
-  void initState() {
-    user.name = widget.username;
-    user.uid = widget.uuid;
-    super.initState();
-  }
-
-  void systemMessage() {
-    Timer(Duration(milliseconds: 300), () {
-      if (i < 6) {
-        setState(() {
-          messages = [...messages, m[i]];
-        });
-        i++;
-      }
-      Timer(Duration(milliseconds: 300), () {
-        _chatViewKey.currentState.scrollController
-          ..animateTo(
-            _chatViewKey.currentState.scrollController.position.maxScrollExtent,
-            curve: Curves.easeOut,
-            duration: const Duration(milliseconds: 300),
-          );
-      });
-    });
-  }
-
-  void onSend(ChatMessage message) {
-    var documentReference = FirebaseFirestore.instance
-        .collection('messages')
-        .doc(DateTime.now().millisecondsSinceEpoch.toString());
-
-    FirebaseFirestore.instance.runTransaction((transaction) async {
-      transaction.set(
-        documentReference,
-        message.toJson(),
-      );
-    });
-  }
-
-  final metadata = firebase_storage.SettableMetadata(
-    contentType: 'image/jpg',
-  );
-
-  void uploadFile() async {
-    PickedFile selectedFile = await ImagePicker().getImage(
-      source: ImageSource.gallery,
-      imageQuality: 80,
-      maxHeight: 400,
-      maxWidth: 400,
-    );
-    File result = File(selectedFile.path);
-
-    if (result != null) {
-      String id = Uuid().v4().toString();
-
-      final firebase_storage.Reference storageRef = firebase_storage
-          .FirebaseStorage.instance
-          .ref()
-          .child("chat_images/$id.jpg");
-
-      firebase_storage.UploadTask uploadTask =
-          storageRef.putFile(result, metadata);
-      firebase_storage.TaskSnapshot download = await uploadTask;
-
-      String url = await download.ref.getDownloadURL();
-
-      ChatMessage message = ChatMessage(text: "", user: user, image: url);
-
-      var documentReference = FirebaseFirestore.instance
-          .collection('messages')
-          .doc(DateTime.now().millisecondsSinceEpoch.toString());
-
-      FirebaseFirestore.instance.runTransaction((transaction) async {
-        transaction.set(
-          documentReference,
-          message.toJson(),
-        );
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("Dash Chat"),
-      ),
-      body: StreamBuilder(
-        stream: FirebaseFirestore.instance.collection('messages').snapshots(),
-        builder: (BuildContext context, AsyncSnapshot<QuerySnapshot> snapshot) {
-          if (snapshot.hasError) {
-            return Text('Something went wrong');
-          } else if (snapshot.connectionState == ConnectionState.waiting) {
-            return Text("Loading");
-          } else if (!snapshot.hasData) {
-            return Center(
-              child: CircularProgressIndicator(
-                valueColor: AlwaysStoppedAnimation<Color>(
-                    Theme.of(context).primaryColor),
-              ),
-            );
-          } else {
-            var messages = snapshot.data.docs
-                .map((DocumentSnapshot document) =>
-                    ChatMessage.fromJson(document.data()))
-                .toList();
-            return DashChat(
-              user: user,
-              messages: messages,
-              inverted: false,
-              sendOnEnter: true,
-              showUserAvatar: false,
-              showAvatarForEveryMessage: false,
-              scrollToBottom: true,
-              onPressAvatar: (ChatUser user) {
-                print("OnPressAvatar: ${user.name}");
-              },
-              onLongPressAvatar: (ChatUser user) {
-                print("OnLongPressAvatar: ${user.name}");
-              },
-              inputMaxLines: 5,
-              messageContainerPadding: EdgeInsets.only(left: 5.0, right: 5.0),
-              alwaysShowSend: true,
-              inputTextStyle: TextStyle(fontSize: 16.0),
-              inputContainerStyle: BoxDecoration(
-                border: Border.all(width: 0.0),
-                color: Colors.white,
-              ),
-              textInputAction: TextInputAction.send,
-              dateFormat: DateFormat('yyyy-MMM-dd'),
-              timeFormat: DateFormat('HH:mm'),
-              inputDecoration: InputDecoration(
-                hintText: "Message here...",
-                border: InputBorder.none,
-              ),
-              onSend: onSend,
-              onQuickReply: (Reply reply) {
-                setState(() {
-                  messages.add(ChatMessage(
-                      text: reply.value,
-                      createdAt: DateTime.now(),
-                      user: user));
-
-                  messages = [...messages];
-                });
-
-                Timer(Duration(milliseconds: 300), () {
-                  _chatViewKey.currentState.scrollController
-                    ..animateTo(
-                      _chatViewKey.currentState.scrollController.position
-                          .maxScrollExtent,
-                      curve: Curves.easeOut,
-                      duration: const Duration(milliseconds: 300),
-                    );
-                  if (i == 0) {
-                    systemMessage();
-                    Timer(Duration(milliseconds: 600), () {
-                      systemMessage();
-                    });
-                  } else {
-                    systemMessage();
-                  }
-                });
-              },
-              onLoadEarlier: () {
-                print("laoding...");
-              },
-              shouldShowLoadEarlier: false,
-              showTraillingBeforeSend: true,
-              trailing: <Widget>[
-                IconButton(
-                  icon: Icon(Icons.photo),
-                  onPressed: uploadFile,
-                )
-              ],
-            );
-          }
-        },
-      ),
-    );
-  }
-}
