+++ lib/main.dart
@@ -1,11 +1,9 @@
 import 'package:firebase_core/firebase_core.dart' as firebase_core;
 import 'package:flutter/material.dart';
-import 'feature/home/home.dart';
-
-import 'feature/chat/chat.dart';
+import 'package:ontime/features/home/home.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await firebase_core.Firebase.initializeApp();
-  runApp(new SearchBarDemoApp());
+  runApp(SearchBarDemoApp());
 }
